cpufreq: cpu-boost: No longer depend on wq freezer and use fb notifier
This way we could trigger the boosts as soon as framebuffer notifies the driver regarding a state change. Because of this however, we need to manually disable the boosts through the fb notifier as well. In order for us to reduce the amount of times framebuffer notifier triggers driver functions, we then create a static boolean that will store the previous blank state to compare to the new blank if there's a change from blank_unblank to another state and vice versa.

Signed-off-by: ederekun <sedrickvince@gmail.com>