sched: Reduce latencies by applying more updated values
Since our devices are far capable than those that exist years ago (the oldest tuning commit dates back 11 years ago), I think there is a need for retuning these values. Thus, thanks to tytydraco and his extensive testing, this commit applies the values at which he currently uses for KTweak as of now.
These values were at first aggressive for me, but then I realized how old the last tuning for these values are which then convinced me that these should be updated for newer and modern processors.
Our devices are far more capable as compared to before when it comes to delivering low latencies without hurting throughput.
Because of this reduced latency, however, we also need to reduce the number of migration as to also reduce interrupt time and not take over numerous scheduling periods.

Signed-off-by: ederekun <sedrickvince@gmail.com>