sched/core: Do not allow hrtick to start at suspend
Since ktime cannot be used whenever timekeeping_suspended is on (it creates a kernel log warning), let's postpone hrtick_start whenever phone is suspended. This way we revert to the HZ timer clock and avoid ktime calling at suspend. This is better as opposed to keeping the generic timekeeping always-on.

Signed-off-by: ederekun <sedrickvince@gmail.com>