sched/dynamic_schedtune: Use HRTIMER for limiting dynamic schedtune's active time frame based on input
It is more efficient to just create an independent HRTIMER than create a whole separated kthread. This will both increase the accuracy of the imposed input time frame as well as significantly decrease consistent load on the system. Along with this change is the simplification of header variables and granting mdss_fb rights to acquire update regardless of given time frame while still ensuring that the core kthread does not go beyond the time frame by adding the condition of time frame state in the wait queue function.

Make sure that waitqueue is running before acquiring an update to ensure that the core kthread will not skip an update and cause untimely updates.

Let us also limit GPU wakeups according to last input by having dynamic schedtune's time frame state as a condition to the execution of KGSL IOCTL GPU wakeup work structure.

Remove sugov_update_util for now to be rewritten in the following next commit to fix issues and for it to be inlined with dynamic_schedtune_set which is located in main schedtune source code.

Signed-off-by: ederekun <sedrickvince@gmail.com>